# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type AccountAccountType =
  (typeof AccountAccountType)[keyof typeof AccountAccountType];

export const AccountAccountType = {
  long_term: "long_term",
  swing: "swing",
  day_trading: "day_trading",
  savings: "savings",
} as const;

export interface Account {
  id: number;
  name: string;
  broker: string;
  accountType: AccountAccountType;
  currency: string;
  initialBalance: number;
  currentBalance: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateAccountBodyAccountType =
  (typeof CreateAccountBodyAccountType)[keyof typeof CreateAccountBodyAccountType];

export const CreateAccountBodyAccountType = {
  long_term: "long_term",
  swing: "swing",
  day_trading: "day_trading",
  savings: "savings",
} as const;

export interface CreateAccountBody {
  name: string;
  broker: string;
  accountType: CreateAccountBodyAccountType;
  currency: string;
  initialBalance: number;
}

export type UpdateAccountBodyAccountType =
  (typeof UpdateAccountBodyAccountType)[keyof typeof UpdateAccountBodyAccountType];

export const UpdateAccountBodyAccountType = {
  long_term: "long_term",
  swing: "swing",
  day_trading: "day_trading",
  savings: "savings",
} as const;

export interface UpdateAccountBody {
  name?: string;
  broker?: string;
  accountType?: UpdateAccountBodyAccountType;
  currentBalance?: number;
}

export interface Position {
  id: number;
  accountId: number;
  symbol: string;
  name: string;
  quantity: number;
  avgCost: number;
  currentPrice: number;
  marketValue: number;
  unrealizedPnl: number;
  unrealizedPnlPct: number;
  sector?: string;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePositionBody {
  accountId: number;
  symbol: string;
  name: string;
  quantity: number;
  avgCost: number;
  sector?: string;
  notes?: string;
}

export interface UpdatePositionBody {
  quantity?: number;
  avgCost?: number;
  currentPrice?: number;
  notes?: string;
}

export type TradeActivityActivityType =
  (typeof TradeActivityActivityType)[keyof typeof TradeActivityActivityType];

export const TradeActivityActivityType = {
  buy: "buy",
  sell: "sell",
  dividend: "dividend",
  deposit: "deposit",
  withdrawal: "withdrawal",
  note: "note",
} as const;

export interface TradeActivity {
  id: number;
  accountId: number;
  symbol?: string;
  activityType: TradeActivityActivityType;
  quantity?: number;
  price?: number;
  totalAmount?: number;
  notes?: string;
  tradeDate: string;
  createdAt: string;
}

export type CreateActivityBodyActivityType =
  (typeof CreateActivityBodyActivityType)[keyof typeof CreateActivityBodyActivityType];

export const CreateActivityBodyActivityType = {
  buy: "buy",
  sell: "sell",
  dividend: "dividend",
  deposit: "deposit",
  withdrawal: "withdrawal",
  note: "note",
} as const;

export interface CreateActivityBody {
  accountId: number;
  symbol?: string;
  activityType: CreateActivityBodyActivityType;
  quantity?: number;
  price?: number;
  totalAmount?: number;
  notes?: string;
  tradeDate: string;
}

export interface MarketQuote {
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  volume?: number;
  marketCap?: number;
  high52w?: number;
  low52w?: number;
  peRatio?: number;
  previousClose?: number;
}

export interface ChartData {
  symbol: string;
  interval: string;
  range: string;
  timestamps: number[];
  opens?: number[];
  highs?: number[];
  lows?: number[];
  closes: number[];
  volumes?: number[];
}

export interface ScreenerResult {
  symbol: string;
  name: string;
  price: number;
  changePercent: number;
  volume: number;
  avgVolume?: number;
  relativeVolume?: number;
  sector?: string;
  marketCap?: number;
  rsi?: number;
  swingScore: number;
}

export interface AccountSummary {
  id: number;
  name: string;
  accountType: string;
  nav: number;
  unrealizedPnl: number;
  unrealizedPnlPct: number;
  positionCount: number;
}

export interface PortfolioSummary {
  totalNav: number;
  totalCost: number;
  totalUnrealizedPnl: number;
  totalUnrealizedPnlPct: number;
  dayChange: number;
  dayChangePct: number;
  accountCount: number;
  positionCount: number;
  accounts: AccountSummary[];
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type ListActivitiesParams = {
  accountId?: number;
  limit?: number;
};

export type GetMarketChartParams = {
  interval?: string;
  range?: string;
};

export type ScreenStocksParams = {
  minPrice?: number;
  maxPrice?: number;
  minVolume?: number;
  sector?: string;
};