# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Portfolio Tracker API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all trading accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new trading account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /accounts/{id}/positions:
    get:
      operationId: listAccountPositions
      tags: [positions]
      summary: List positions for an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List all trade activities
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeActivity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Log a trade activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityBody"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeActivity"
  /activities/{id}:
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /market/quote/{symbol}:
    get:
      operationId: getMarketQuote
      tags: [market]
      summary: Get stock quote
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketQuote"
        "404":
          description: Symbol not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /market/chart/{symbol}:
    get:
      operationId: getMarketChart
      tags: [market]
      summary: Get stock chart data
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: interval
          in: query
          schema:
            type: string
            default: "1d"
        - name: range
          in: query
          schema:
            type: string
            default: "1mo"
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartData"
...
  /market/screener:
    get:
      operationId: screenStocks
      tags: [market]
      summary: Screen stocks for swing trading
      parameters:
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minVolume
          in: query
          schema:
            type: number
        - name: sector
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of screened stocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScreenerResult"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        broker:
          type: string
        accountType:
          type: string
          enum: [long_term, swing, day_trading, savings]
        currency:
          type: string
        initialBalance:
          type: number
        currentBalance:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - broker
        - accountType
        - currency
        - initialBalance
        - currentBalance
        - createdAt
        - updatedAt
    CreateAccountBody:
      type: object
      properties:
        name:
          type: string
        broker:
          type: string
        accountType:
          type: string
          enum: [long_term, swing, day_trading, savings]
        currency:
          type: string
        initialBalance:
          type: number
      required:
        - name
        - broker
        - accountType
        - currency
        - initialBalance
    UpdateAccountBody:
      type: object
      properties:
        name:
          type: string
        broker:
          type: string
        accountType:
          type: string
          enum: [long_term, swing, day_trading, savings]
        currentBalance:
          type: number
    Position:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        symbol:
          type: string
        name:
          type: string
        quantity:
          type: number
        avgCost:
          type: number
        currentPrice:
          type: number
        marketValue:
          type: number
        unrealizedPnl:
          type: number
        unrealizedPnlPct:
          type: number
        sector:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - accountId
        - symbol
        - name
        - quantity
        - avgCost
        - currentPrice
        - marketValue
        - unrealizedPnl
        - unrealizedPnlPct
        - createdAt
        - updatedAt
...
    TradeActivity:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        symbol:
          type: string
        activityType:
          type: string
          enum: [buy, sell, dividend, deposit, withdrawal, note]
        quantity:
          type: number
        price:
          type: number
        totalAmount:
          type: number
        notes:
          type: string
        tradeDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - accountId
        - activityType
        - tradeDate
        - createdAt
    CreateActivityBody:
      type: object
      properties:
        accountId:
          type: integer
        symbol:
          type: string
        activityType:
          type: string
          enum: [buy, sell, dividend, deposit, withdrawal, note]
        quantity:
          type: number
        price:
          type: number
        totalAmount:
          type: number
        notes:
          type: string
        tradeDate:
          type: string
          format: date-time
      required:
        - accountId
        - activityType
        - tradeDate
    MarketQuote:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        volume:
          type: number
        marketCap:
          type: number
        high52w:
          type: number
        low52w:
          type: number
        peRatio:
          type: number
        previousClose:
          type: number
      required:
        - symbol
        - name
        - price
        - change
        - changePercent
    ChartData:
      type: object
      properties:
        symbol:
          type: string
        interval:
          type: string
        range:
          type: string
        timestamps:
          type: array
          items:
            type: integer
        opens:
          type: array
          items:
            type: number
        highs:
          type: array
          items:
            type: number
        lows:
          type: array
          items:
            type: number
        closes:
          type: array
          items:
            type: number
        volumes:
          type: array
          items:
            type: number
      required:
        - symbol
        - interval
        - range
        - timestamps
        - closes
    ScreenerResult:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        price:
          type: number
        changePercent:
          type: number
        volume:
          type: number
        avgVolume:
          type: number
        relativeVolume:
          type: number
        sector:
          type: string
        marketCap:
          type: number
        rsi:
          type: number
        swingScore:
          type: number
      required:
        - symbol
        - name
        - price
        - changePercent
        - volume
        - swingScore
...