# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccountAccountType } from "./accountAccountType";

export interface Account {
  id: number;
  name: string;
  broker: string;
  accountType: AccountAccountType;
  currency: string;
  initialBalance: number;
  currentBalance: number;
  createdAt: Date;
  updatedAt: Date;
}