# lib/api-zod/src/generated/types/createActivityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateActivityBodyActivityType } from "./createActivityBodyActivityType";

export interface CreateActivityBody {
  accountId: number;
  symbol?: string;
  activityType: CreateActivityBodyActivityType;
  quantity?: number;
  price?: number;
  totalAmount?: number;
  notes?: string;
  tradeDate: Date;
}