# lib/api-zod/src/generated/types/screenerResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScreenerResult {
  symbol: string;
  name: string;
  price: number;
  changePercent: number;
  volume: number;
  avgVolume?: number;
  relativeVolume?: number;
  sector?: string;
  marketCap?: number;
  rsi?: number;
  swingScore: number;
}