# lib/api-zod/src/generated/types/tradeActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeActivityActivityType } from "./tradeActivityActivityType";

export interface TradeActivity {
  id: number;
  accountId: number;
  symbol?: string;
  activityType: TradeActivityActivityType;
  quantity?: number;
  price?: number;
  totalAmount?: number;
  notes?: string;
  tradeDate: Date;
  createdAt: Date;
}