# lib/api-zod/src/generated/types/updateAccountBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAccountBodyAccountType } from "./updateAccountBodyAccountType";

export interface UpdateAccountBody {
  name?: string;
  broker?: string;
  accountType?: UpdateAccountBodyAccountType;
  currentBalance?: number;
}